+++ admin/tabs/clients/ClientsModel.ts
@@ -81,19 +81,18 @@ export class ClientsModel extends BaseAdminTabModel {
             .newSpan('load')
             .run(async ctx => {
                 const data = await ctx.fetchJson({url: 'clientAdmin/allClients'});
-                if (loadSpec.isStale) return;
 
                 gridModel.loadData(data);
                 gridModel.preSelectFirstAsync();
                 runInAction(() => {
                     this.lastRefresh = Date.now();
                 });
-            })
-            .catch(e => {
-                if (loadSpec.isStale || loadSpec.isAutoRefresh) return;
-                gridModel.clear();
-                XH.handleException(e, {alertType: 'toast'});
             });
+    }
+
+    override handleLoadException(e: unknown) {
+        this.gridModel.clear();
+        XH.handleException(e, {alertType: 'toast'});
     }
 
     async forceSuspendAsync(toRecs: StoreRecord[]) {