+++ cmp/viewmanager/ViewManagerModel.ts
@@ -363,12 +363,11 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
 
     override async doLoadAsync(loadSpec: LoadSpec) {
         const {dataAccess, view} = this;
-        await this.runOn(loadSpec)
+        return this.runOn(loadSpec)
             .newSpan('refresh')
             .run(async ctx => {
                 // 1) Update views and related state
                 const {views, state} = await dataAccess.fetchDataAsync(ctx);
-                if (loadSpec.isStale) return;
                 runInAction(() => {
                     this.views = views;
                     this.userPinned = state.userPinned;
@@ -382,11 +381,11 @@ export class ViewManagerModel<T = PlainObject> extends HoistModel {
                         this.loadViewAsync(view.token, this.pendingValue, ctx);
                     }
                 }
-            })
-            .catch(e => {
-                if (loadSpec.isStale) return;
-                this.handleException(e, {showAlert: false});
             });
+    }
+
+    override handleLoadException(e: unknown) {
+        this.handleException(e, {showAlert: false});
     }
 
     async selectViewAsync(